app/requirements.txt
Flask
flask_sqlalchemy
flask_wtf
pymysql
cryptography
python-dotenv